Fixed error : call of overloaded 'noTone(<anonymous enum>)' is ambiguous (#94)

+++ k3ng_keyer/keyer_stm32duino.h
@@ -34,6 +34,6 @@
   void service_send_buffer(byte);
   void check_ptt_tail(void);
   void check_for_dirty_configuration(void);
-  void tone(uint8_t, short unsigned int, unsigned int);
-  void noTone(uint8_t);
+  void tone(uint32_t, uint32_t, uint32_t);
+  void noTone(uint32_t);
   void serial_status(PRIMARY_SERIAL_CLS*);
